added color_histogramm to the db
